Day 1: Sonar Sweep
Calculate the number of times a sequence of integers increases, and then calculate the same, but for
a summed rolling window of three across the input dataset.

   python3 sonar_sweep.py

To my astonishment, this worked on the first attempt. Ideally I would not be using index arithmetic
to calculate the summed rolling window. There has to be a more Pythonic way of accomplishing the
same thing.