Day 2: Dive!
With a list of instructions, with each line consisting of one of `forward`,
`up` or `down` followed by an integer representing the magnitude, calculate the
total depth and horizontal position of your submarine (remember that down and
up refer to depth!). Multiply the horizontal positio and depth for the first
answer.

   python3 depth.py

For the second question, the logic is changed: `down` and `up` commands affect
a parameter called `aim`. The `forward` command then increases the horizontal
position and independently affects the depth by a multiple of the calculated
`aim` variable.

A classic switch statement in Python, despite it not supporting such magic.
There are far more clever ways of doing this, however a simple conditional
statement is clearer, and the dataset is sufficiently small for this logic to
be optimal.